uploaded docs(reports) and excel files (collected data)
I started the project in June, I have provided the background information, the data I collected for the research as well as a cleaned and summarized version of it for the first month. I provided a report for that as well in the repo.

The main dashboard and charts will be provided at the end of this month, after the second part of the research is completed